+++ src/RecoveryServices/RecoveryServices/ChangeLog.md
@@ -18,6 +18,7 @@
         - Additional information about change #1
 -->
 ## Upcoming Release
+* Fixed delay in long running operations [#18567]
 
 ## Version 5.4.0
 * Added support for Multi-user authorization using Resource Guard for recovery services vault.